# openAPI.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Good Life Financial Services",
    "version": "1.7.0",
    "license": {
      "name": "MIT",
      "url": "https://github.com/Na5co/devtal"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000/v1",
      "description": "Development server"
    },
    {
      "url": "https://dev-portal-7xkr.onrender.com/v1",
      "description": "Production server"
    }
  ],
  "paths": {
    "/admin/fraud-status/{identifier}": {
      "post": {
        "summary": "Set fraud status for a user",
        "tags": [
          "Admin"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "identifier",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "User ID or Email"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "status": {
                    "type": "string",
                    "enum": [
                      "fraud",
                      "not_fraud"
                    ]
                  }
                },
                "example": {
                  "status": "fraud"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
    "/admin/credit-score/{identifier}": {
      "post": {
        "summary": "Set credit score for a user",
        "tags": [
          "Admin"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "identifier",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "User ID or Email"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "score": {
                    "type": "number"
                  }
                },
                "example": {
                  "score": 750
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
    "/admin/loan-application/{identifier}/review": {
      "patch": {
        "summary": "Review a loan application",
        "description": "Allows an admin to approve or decline a loan application that is pending review.",
        "tags": [
          "Admin"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "identifier",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "User ID or Email"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "decision": {
                    "type": "string",
                    "enum": [
                      "approved",
                      "declined"
                    ],
                    "description": "The decision on the loan application."
                  }
                },
                "example": {
                  "decision": "approved"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK. The loan application status has been updated."
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
    "/admin/risk-assessment/{identifier}": {
      "post": {
        "summary": "Assess the financial risk of a user",
        "description": "Allows an admin to trigger a financial risk assessment for a specific user. This must be done before a user can apply for a loan.",
        "tags": [
          "Admin"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "identifier",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "User ID or Email"
          }
        ],
        "responses": {
          "200": {
            "description": "OK. The risk assessment has been completed.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "userId": {
                      "type": "string"
                    },
                    "riskLevel": {
                      "type": "string",
                      "enum": [
                        "low",
                        "medium",
                        "high"
                      ]
                    },
                    "recommendation": {
                      "type": "string",
                      "enum": [
                        "proceed",
                        "proceed_with_caution",
                        "deny"
                      ]
                    },
                    "assessment": {
                      "type": "object",
                      "properties": {
                        "creditScoreCheck": {
                          "type": "string"
                        },
                        "dtiRatioCheck": {
                          "type": "string"
                        },
                        "fraudFlags": {
                          "type": "array",
                          "items": {
                            "type": "string"
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
    "/admin/loan-applications": {
      "get": {
        "summary": "Get all loan applications",
        "description": "Allows an admin to retrieve all loan applications, with options for filtering and pagination.",
        "tags": [
          "Admin"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "status",
            "schema": {
              "type": "string",
              "enum": [
                "pending_review",
                "approved",
                "declined"
              ]
            },
            "description": "Filter applications by their status."
          },
          {
            "in": "query",
            "name": "sortBy",
            "schema": {
              "type": "string"
            },
            "description": "Sort by query in the form of field:desc/asc (ex. name:asc)."
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer",
              "minimum": 1
            },
            "description": "Maximum number of applications."
          },
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "integer",
              "minimum": 1
            },
            "description": "Page number."
          }
        ],
        "responses": {
          "200": {
            "description": "OK. A list of loan applications."
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          }
        }
      }
    },
    "/admin/assessments": {
      "get": {
        "summary": "Get all risk assessments",
        "description": "Allows an admin to retrieve all risk assessments, with options for filtering and pagination.",
        "tags": [
          "Admin"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "status",
            "schema": {
              "type": "string",
              "enum": [
                "pending_assessment",
                "pending_review",
                "approved",
                "declined"
              ]
            },
            "description": "Filter assessments by their status."
          },
          {
            "in": "query",
            "name": "sortBy",
            "schema": {
              "type": "string"
            },
            "description": "Sort by query in the form of field:desc/asc (ex. name:asc)."
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer",
              "minimum": 1
            },
            "description": "Maximum number of assessments."
          },
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "integer",
              "minimum": 1
            },
            "description": "Page number."
          }
        ],
        "responses": {
          "200": {
            "description": "OK. A list of risk assessments."
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          }
        }
      }
    },
    "/admin/loans/active": {
      "get": {
        "summary": "Get all active loans",
        "description": "Allows an admin to retrieve all active loans, with options for pagination.",
        "tags": [
          "Admin"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "sortBy",
            "schema": {
              "type": "string"
            },
            "description": "Sort by query in the form of field:desc/asc (ex. name:asc)."
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer",
              "minimum": 1
            },
            "description": "Maximum number of loans."
          },
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "integer",
              "minimum": 1
            },
            "description": "Page number."
          }
        ],
        "responses": {
          "200": {
            "description": "OK. A list of active loans."
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Error": {
        "type": "object",
        "properties": {
          "code": {
            "type": "number"
          },
          "message": {
            "type": "string"
          }
        }
      }
    },
    "responses": {
...
      "Unauthorized": {
        "description": "Unauthorized",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            },
            "example": {
              "code": 401,
              "message": "Please authenticate"
            }
          }
        }
      },
      "Forbidden": {
        "description": "Forbidden",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            },
            "example": {
              "code": 403,
              "message": "Forbidden"
            }
          }
        }
      },
      "NotFound": {
        "description": "Not found",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            },
            "example": {
              "code": 404,
              "message": "Not found"
            }
          }
        }
      }
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
...
}